--- low-level/files/opt/seagate/sspl/setup/health_provider/error.py
@@ -1,27 +1,0 @@
-#!/bin/env python3
-
-# CORTX Python common library.
-# Copyright (c) 2021 Seagate Technology LLC and/or its Affiliates
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as published
-# by the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU Affero General Public License for more details.
-# You should have received a copy of the GNU Affero General Public License
-# along with this program. If not, see <https://www.gnu.org/licenses/>.
-# For any questions about this software or licensing,
-# please email opensource@seagate.com or cortx-questions@seagate.com.
-
-from cortx.utils.errors import BaseError
-
-
-class HealthProviderError(BaseError):
-    """Generic Exception with error code and output."""
-
-    def __init__(self, rc=0, desc=None, message_id=None, message_args=None):
-        """Initialize HealthProviderError"""
-        super(HealthProviderError, self).__init__(
-              rc, 'HealthProvider Error: %s' % desc, message_id, message_args)
