--- low-level/files/opt/seagate/sspl/setup/health_provider/server.py
@@ -1,40 +1,0 @@
-#!/bin/python3
-
-# CORTX Python common library.
-# Copyright (c) 2021 Seagate Technology LLC and/or its Affiliates
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as published
-# by the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU Affero General Public License for more details.
-# You should have received a copy of the GNU Affero General Public License
-# along with this program. If not, see <https://www.gnu.org/licenses/>.
-# For any questions about this software or licensing,
-# please email opensource@seagate.com or cortx-questions@seagate.com
-
-"""
- ***************************************************************************
-  Description: Server class to provide fru health and specific information
- ***************************************************************************
-"""
-
-from resource_health import ResourceHealth
-
-
-class ServerHealth(ResourceHealth):
-    """Provides health information of FRUs in a node"""
-
-    name = "server"
-
-    def __init__(self):
-        """Initialize server"""
-        pass
-
-    def get_health_info(self, rid):
-        """Get health information of fru in given resource id
-        rid: Resource id (Example: hw>disks)
-        """
-        return
